+++ gui/Pages/FluidCharacteristicsModel.py
@@ -151,32 +151,25 @@ class FluidCharacteristicsModel(Variables, Model):
         (also called by NumericalParamGlobalView and TimeStepView)
         """
         d = {}
-        d['joule_effect']       = 'off'
-        d['gas_combustion']     = 'off'
-        d['solid_fuels']        = 'off'
-        d['thermal_scalar']     = 'off'
-        d['atmospheric_flows']  = 'off'
-        d['compressible_model'] = 'off'
 
-        node1 = self.node_models.xmlGetNode('gas_combustion',    'model')
-        node2 = self.node_models.xmlGetNode('solid_fuels',       'model')
-        node3 = self.node_models.xmlGetNode('joule_effect',      'model')
-        node4 = self.node_models.xmlGetNode('thermal_scalar',    'model')
-        node5 = self.node_models.xmlGetNode('atmospheric_flows', 'model')
-        node6 = self.node_models.xmlGetNode('compressible_model', 'model')
+        for mdl in ['atmospheric_flows', 'compressible_model',
+                    'gas_combustion', 'joule_effect', 'solid_fuels',
+                    'thermal_scalar']:
+            d[mdl] = 'off'
 
-        for node in (node1, node2, node3, node4, node5, node6):
+            node = self.node_models.xmlGetNode(mdl, 'model')
+
             if node:
                 if node['model'] == "":
                     node['model'] = "off"
                 if node['model'] != 'off':
-                    d[node.el.tagName] = node['model']
+                    d[mdl] = node['model']
 
         return d['atmospheric_flows'], \
                d['joule_effect'],      \
                d['thermal_scalar'],    \
                d['gas_combustion'],    \
-               d['solid_fuels'],    \
+               d['solid_fuels'],       \
                d['compressible_model']