+++ NEWS.md
@@ -3,6 +3,8 @@ Master (not on release branches yet)
 
 User changes:
 
+- Allow stopping criteria based on physical time and/or additional time.
+
 - Add cs_restart_map_set_mesh_input function to allow mapping restarts
   from computation using a different mesh.